+++ circuits/updater.circom
@@ -1,0 +1,4 @@
+include "./update_state_tree.circom"
+
+
+component main = UpdateStateTree(4, 2);
